+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/CombineLatest6Suite.scala
@@ -32,7 +32,7 @@ object CombineLatest6Suite extends BaseOperatorSuite {
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
     val o5 = Observable.now(5)
-    val o6 = Observable.range(0, sourceCount)
+    val o6 = Observable.range(0L, sourceCount.toLong)
     val o = Observable.combineLatestMap6(o1, o2, o3, o4, o5, o6)(_ + _ + _ + _ + _ + _)
 
     Sample(o, count(sourceCount), sum(sourceCount), waitFirst, waitNext)
@@ -49,7 +49,7 @@ object CombineLatest6Suite extends BaseOperatorSuite {
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
     val o5 = Observable.now(5)
-    val flawed = createObservableEndingInError(Observable.range(0, sourceCount), ex)
+    val flawed = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
     val o = Observable.combineLatestMap6(o1, o2, o3, o4, o5, flawed)(_ + _ + _ + _ + _ + _)
 
     Sample(o, count(sourceCount - 1), sum(sourceCount - 1), waitFirst, waitNext)
@@ -62,13 +62,13 @@ object CombineLatest6Suite extends BaseOperatorSuite {
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
     val o5 = Observable.now(5)
-    val o6 = Observable.range(0, sourceCount)
+    val o6 = Observable.range(0L, sourceCount.toLong)
 
     val o = Observable.combineLatestMap6(o1, o2, o3, o4, o5, o6) { (a1, a2, a3, a4, a5, a6) =>
       if (a6 == sourceCount - 1) throw dummy else a1 + a2 + a3 + a4 + a5 + a6
     }
 
-    Sample(o, count(sourceCount - 1), sum(sourceCount - 1), waitFirst, waitNext)
+    Sample(o, count(sourceCount - 1), sum(sourceCount - 1).toLong, waitFirst, waitNext)
   }
 
   override def cancelableObservables(): Seq[Sample] = {
