+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/EchoRepeatedSuite.scala
@@ -30,8 +30,8 @@ object EchoRepeatedSuite extends BaseOperatorSuite {
   def waitNext = 1.second
 
   def createObservable(sourceCount: Int) = Some {
-    val source = Observable.now(1L).delayOnComplete(sourceCount.seconds * 2)
-    val o = source.echoRepeated(1.second).drop(1).take(sourceCount)
+    val source = Observable.now(1L).delayOnComplete(sourceCount.seconds * 2.toLong)
+    val o = source.echoRepeated(1.second).drop(1).take(sourceCount.toLong)
     Sample(o, count(sourceCount), sum(sourceCount), waitFirst, waitNext)
   }
