+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/Zip3Suite.scala
@@ -23,9 +23,9 @@ import scala.concurrent.duration.Duration.Zero
 
 object Zip3Suite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Some {
-    val o1 = Observable.range(0, sourceCount).executeAsync
-    val o2 = Observable.range(0, sourceCount + 1).executeAsync
-    val o3 = Observable.range(0, sourceCount + 2).executeAsync
+    val o1 = Observable.range(0L, sourceCount.toLong).executeAsync
+    val o2 = Observable.range(0, sourceCount.toLong + 1).executeAsync
+    val o3 = Observable.range(0, sourceCount.toLong + 2).executeAsync
 
     val o = Observable.zipMap3(o1, o2, o3)(_ + _ + _)
     Sample(o, count(sourceCount), sum(sourceCount), Zero, Zero)
@@ -35,18 +35,18 @@ object Zip3Suite extends BaseOperatorSuite {
   def sum(sourceCount: Int) = (sourceCount * (sourceCount - 1)) / 2 * 3
 
   def observableInError(sourceCount: Int, ex: Throwable) = Some {
-    val o1 = createObservableEndingInError(Observable.range(0, sourceCount), ex)
-    val o2 = createObservableEndingInError(Observable.range(0, sourceCount), ex)
-    val o3 = createObservableEndingInError(Observable.range(0, sourceCount), ex)
+    val o1 = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
+    val o2 = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
+    val o3 = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
 
     val o = Observable.zipMap3(o1, o2, o3)(_ + _ + _)
     Sample(o, count(sourceCount - 1), sum(sourceCount - 1), Zero, Zero)
   }
 
   def brokenUserCodeObservable(sourceCount: Int, ex: Throwable) = Some {
-    val o1 = Observable.range(0, sourceCount)
-    val o2 = Observable.range(0, sourceCount + 100)
-    val o3 = Observable.range(0, sourceCount)
+    val o1 = Observable.range(0L, sourceCount.toLong)
+    val o2 = Observable.range(0, sourceCount.toLong + 100)
+    val o3 = Observable.range(0L, sourceCount.toLong)
 
     val o = Observable.zipMap3(o1, o2, o3) { (x1, x2, x3) =>
       if (x2 < sourceCount - 1) x1 + x2 + x3 else throw ex
